refactor: use constant, remove unused code
