+++ apps/calendar/src/components/view/day.tsx
@@ -8,7 +8,7 @@ import { capitalizeDayName, getDayName } from '@src/util/dayName';
 import { DayGridEvents } from '@src/components/panelgrid/dayGridEvents';
 import TZDate from '@src/time/date';
 import { Layout } from '@src/components/layout';
-import { getDayGridEvents } from '@src/util/gridEvent';
+import { getDayGridEvents } from '@src/util/gridHelper';
 import { toEndOfDay, toStartOfDay } from '@src/time/datetime';
 import { TimeGrid } from '@src/components/timegrid/timegrid';
 import { ColumnInfo } from '@src/components/timegrid/columns';
