+++ apps/calendar/src/components/view/week.tsx
@@ -8,7 +8,7 @@ import { TemplateWeekDay } from '@src/model';
 import { capitalizeDayName, getDayName } from '@src/util/dayName';
 import TZDate from '@src/time/date';
 import { Layout } from '@src/components/layout';
-import { getDayGridEvents } from '@src/util/gridEvent';
+import { getDayGridEvents } from '@src/util/gridHelper';
 import { addDate, toEndOfDay, toStartOfDay } from '@src/time/datetime';
 import { TimeGrid } from '@src/components/timegrid/timegrid';
 import { DayGridEvents } from '@src/components/panelgrid/dayGridEvents';
