+++ apps/calendar/src/util/gridHelper.ts
@@ -1,4 +1,5 @@
 import { findByDateRange } from '@src/controller/month';
+import { findByDateRange as findByDateRangeForWeek } from '@src/controller/week';
 import { DataStore } from '@src/model';
 import ScheduleViewModel from '@src/model/scheduleViewModel';
 import TZDate from '@src/time/date';
@@ -9,8 +10,9 @@ import {
   toStartOfDay,
   withinRangeDate,
 } from '@src/time/datetime';
-import { Cells } from '@t/panel';
+import { Cells, Panel } from '@t/panel';
 import { findIndex, isNil } from '@src/util/utils';
+import { DayGridEventMatrix, EventModelMap, TimeGridEventMatrix } from '@t/events';
 
 export const EVENT_HEIGHT = 22;
 export const TOTAL_WIDTH = 100;
@@ -76,7 +78,7 @@ export function getWidth(widthList: number[], start: number, end: number) {
   }, 0);
 }
 
-const isInGrid = (gridDate: TZDate) => {
+export const isInGrid = (gridDate: TZDate) => {
   return (viewModel: ScheduleViewModel) => {
     const scheduleStart = toStartOfDay(viewModel.getStarts());
     const scheduleEnd = toStartOfDay(viewModel.getEnds());
@@ -188,3 +190,122 @@ export function getRenderedEventViewModels(
     gridDateEventModelMap,
   };
 }
+
+const getDayGridEventModels = (
+  eventModels: DayGridEventMatrix,
+  cells: Cells,
+  narrowWeekend = false
+): DayGridEventMatrix => {
+  eventModels.forEach((matrix) => {
+    matrix.forEach((row) => {
+      row.forEach((viewModel) => {
+        const modelStart = viewModel.getStarts();
+        const modelEnd = viewModel.getEnds();
+        const { width, left } = getEventLeftAndWidth(modelStart, modelEnd, cells, narrowWeekend);
+
+        viewModel.width = width;
+        viewModel.left = left;
+      });
+    });
+  });
+
+  return eventModels;
+};
+
+const getModels = (models: ScheduleViewModel[]) => models.filter((model) => !!model);
+
+export const flattenMatrix = (matrices: DayGridEventMatrix): ScheduleViewModel[] => {
+  const eventModels: ScheduleViewModel[] = [];
+
+  matrices.forEach((matrix) => {
+    matrix.forEach((models) => {
+      eventModels.push(...getModels(models));
+    });
+  });
+
+  return eventModels;
+};
+
+export const setDayGridEventModels = (models: ScheduleViewModel[]): ScheduleViewModel[] => {
+  models.forEach((model) => {
+    model.top += 1;
+  });
+
+  return models;
+};
+
+const getTimeGridEventModels = (
+  eventModels: TimeGridEventMatrix,
+  cells: Cells,
+  narrowWeekend = false
+): ScheduleViewModel[] => {
+  const result: ScheduleViewModel[] = [];
+
+  Object.values(eventModels).forEach((matrices) =>
+    matrices.forEach((matrix) => {
+      matrix.forEach((row) => {
+        result.push(...getModels(row));
+      });
+    })
+  );
+
+  return result;
+};
+
+export const getDayGridEvents = (
+  cells: Cells,
+  dataStore: DataStore,
+  narrowWeekend: boolean
+): EventModelMap => {
+  const panels = [
+    {
+      name: 'milestone',
+      type: 'daygrid',
+      show: true,
+    },
+    {
+      name: 'task',
+      type: 'daygrid',
+      show: true,
+    },
+    {
+      name: 'allday',
+      type: 'daygrid',
+      show: true,
+    },
+    {
+      name: 'time',
+      type: 'timegrid',
+      show: true,
+    },
+  ] as Panel[];
+  const eventModels = findByDateRangeForWeek(dataStore, {
+    start: toStartOfDay(cells[0]),
+    end: toEndOfDay(cells[cells.length - 1]),
+    panels,
+    andFilters: [],
+    options: {
+      hourStart: 0,
+      hourEnd: 24,
+    },
+  });
+
+  return Object.keys(eventModels).reduce<EventModelMap>(
+    (acc, cur) => {
+      const events = eventModels[cur as keyof EventModelMap];
+
+      return {
+        ...acc,
+        [cur]: Array.isArray(events)
+          ? getDayGridEventModels(events, cells, narrowWeekend)
+          : getTimeGridEventModels(events, cells, narrowWeekend),
+      };
+    },
+    {
+      milestone: [],
+      allday: [],
+      task: [],
+      time: [],
+    }
+  );
+};
